Rewrite with @panzi's fork, no more submodule
Panzi's SocialSharePrivacy fork has a build.sh that generates
a compact distribution (this is where you define which share button
modules you want to include - if you want to change that).

sharefest/public/SocialSharePrivacy/ is a result of such a build,
and contains a README explaining how to regenerate it.
This is not a git submodule in order to simplify zip installation,
and minimize zip size.